2022 Update
This commit brings the project to 2022 in the eyes of node.js development. All dependencies were updated to their latest versions and the code was modified to suit breaking changes that these new versions brought:

The Serial Port package brought some small changes to how the package was included (line 42) and to how a new serial port was created (line 50) where now all data about the port is passed through an options array.

On the other hand, the update in socket.io just required also updating the library included in the visualization, so that the version matched (index.html).

Some other changes were done to the server code: changed the serial port path of the coordinator (has to be done if coordinator device changes), changed the lifxMAC values (just for tests, in the future we will include all of them), added a call to serverClock (line 92) to run instantly when the server starts, and commented out a console log (line 172) that visualized current flow control out to the XBees. In the future, the idea is to have a debug flag that can be turned on or off if we want to see those logs or not.