Updated code
The project was updated and brought forward to the most recent node.js version. All dependencies were updated as well.

Arduino code was updated to reflect the new XBee being used as coordinator. This is a crucial change that has to be done every time we change of coordinator. Furthermore, some functions in the server were modified to accomodate using 6 bulbs now. These were also changed to animate the transitions between the bulbs more smoothly.