Unify controller code
Previously, each of the 14 controllers had its own arduino sketch. All of these were pretty much the same, except that we had to manually write the address of the connected XBee in each one. This commit updated the code so that this address is now not manually written but rather requested from the XBee on initialization. This now allows all controllers to have the same exact code.

A small caveat is that, these different sketches also had slightly different calibration values for the measurable range of the potentiometers, as each one is a little different. For now, the code is unified, but we may need to find a solution for this if it becomes a problem.