Update lifx node dependency
The project was developed using the node-lifx dependency. However, it was more-or-less recently deprecated and has not received any updates in the last 7 years. Therefore, this commit replaces this dependency with lifx-lan-client, a fork of the original project that is said to be actively maintained and received its latest update 6 months ago at the time of this commit. A few short tests show that the existing code is compatible with this dependency, as they have done an effort to maintain the forks methods.