Better order for xbee message handler
Before this commit, the function that handled all incoming xbee frames was attached to the message received event in the general synchronous part of the code. I suppose this is the root of errors being generated when the program receives a message just a few seconds after having started. Therefore, the initiation of this function has been moved inside the serial-available code block, so it only runs once serial communication with the xbee has been rightfully established. This fix should remain in observation, as it has not yet been fully proven that this was the root of said problem.